Fixed a bug where redundant org.panini.runtime.* and duck imports were being generated. This was due to a block of code which was being executed on each call to visitCapsuleDef.
This block was shuffled off to a method named
setRuntimeImports(Env<AttrContext> env)
and is now called on topEnv in the method visitTopLevel. This asserts
that the block is executed once per compilation unit. It also fixes the
issue where no runtime imports were generated for compilation units
containing only a single system declaration.