The class org.paninij.lang.Long is now a Duck.
Implemented the Panini$Duck interface, with java.lang.Long as the
generic type.

Did not modify the 'static' helper that return primitives. It may be the
case that these helpers should be deleted instead of duplicated
logic/code from the JRE.

Includes a basic junit smoke test that creating an org.paninij.lang.Long
properly sets the redeemed flag when constructed with or without the
final value, and that the panini$finish method works correctly.