Removed case for CAPSULE_WIRING visitIdentifier and checkId.
I don't think the special case logic is needed. We want id's to just
come back as their capsule type, and from there we can get the wiring
information. Was specially creating the wiring templates ala method
invocations, but wiring is simpler and we can get rid of some of the
cruft.