Changed the Syntax and Semantics of "Associate"
Associate is now structured in a manner similar
to System.arraycopy, that is, the keyword
"associate" now requires a minimum of five arguments
which represent: 1) the source capsules array representing
the capsules being wired, 2) the source position to begin
the wiring, 3) the destination capsule array representing
the capsules to which to source is being wired, 4) the starting
position in the destination capsule array, 5) the number of
elements to be wired. All other further arguments are treated
as additional arguments that will be used in the capsule wiring.